+++ src/main/java/com/example/gfup2/domain/model/RefreshToken.java
@@ -1,0 +1,37 @@
+package com.example.gfup2.domain.model;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Entity
+@NoArgsConstructor
+@AllArgsConstructor
+public class RefreshToken {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(columnDefinition = "text")
+    private String refreshToken;
+
+    @Column(unique = true)
+    private String accountEmail;
+
+    public RefreshToken(String refreshToken, String emailId) {
+        this.refreshToken = refreshToken;
+        this.accountEmail = emailId;
+    }
+
+    public RefreshToken updateToken(String token) {
+        this.refreshToken = token;
+        return this;
+    }
+
+}
