+++ src/main/java/com/example/gfup2/domain/model/UserRoleEnum.java
@@ -1,0 +1,6 @@
+package com.example.gfup2.domain.model;
+
+public enum UserRoleEnum {
+    USER,
+    ADMIN;
+}
