+++ src/main/java/com/example/gfup2/domain/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package com.example.gfup2.domain.repository;
+
+
+import com.example.gfup2.domain.model.User;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByEmailId(String emailId);
+
+}
