--- src/main/java/com/example/gfup2/domain/user/UserDetailsImpl.java
@@ -1,61 +1,0 @@
-package com.example.gfup2.domain.user;
-
-import com.example.gfup2.domain.user.entity.User;
-import java.util.ArrayList;
-import java.util.Collection;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-public class UserDetailsImpl implements UserDetails {
-
-    private User user;
-    private Collection<? extends GrantedAuthority> authorities;
-
-    public static UserDetailsImpl from(User user){
-        SimpleGrantedAuthority simpleGrantedAuthority = new SimpleGrantedAuthority(user.getRole().toString());
-        Collection<GrantedAuthority> collection = new ArrayList<>();
-        collection.add(simpleGrantedAuthority);
-
-        UserDetailsImpl userDetails = new UserDetailsImpl();
-        userDetails.user = user;
-        userDetails.authorities = collection;
-
-        return userDetails;
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return authorities;
-    }
-
-    @Override
-    public String getPassword() {
-        return user.getPassword();
-    }
-
-    @Override
-    public String getUsername() {
-        return user.getEmailId();
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return false;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return false;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return false;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return false;
-    }
-}
