--- src/main/java/com/example/gfup2/domain/user/entity/UserRoleEnum.java
@@ -1,6 +1,0 @@
-package com.example.gfup2.domain.user.entity;
-
-public enum UserRoleEnum {
-    USER,
-    ADMIN;
-}
