--- src/main/java/com/example/gfup2/domain/user/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package com.example.gfup2.domain.user.repository;
-
-
-import com.example.gfup2.domain.user.entity.User;
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmailId(String emailId);
-
-}
