+++ src/main/java/com/example/gfup2/dto/MessageDto.java
@@ -1,0 +1,10 @@
+package com.example.gfup2.dto;
+
+import lombok.*;
+
+@NoArgsConstructor
+@Setter
+@Getter
+public class MessageDto {
+    String to;
+}
