+++ src/main/java/com/example/gfup2/dto/NcpResponseDto.java
@@ -1,0 +1,12 @@
+package com.example.gfup2.dto;
+
+import lombok.*;
+import java.time.LocalDateTime;
+
+@NoArgsConstructor
+public class NcpResponseDto {
+    String requestId;
+    LocalDateTime requestTime;
+    String statusCode;
+    String statusName;
+}
