--- src/main/java/com/example/gfup2/web/verification/dto/MessageDto.java
@@ -1,10 +1,0 @@
-package com.example.gfup2.web.verification.dto;
-
-import lombok.*;
-
-@NoArgsConstructor
-@Setter
-@Getter
-public class MessageDto {
-    String to;
-}
